elements needed for development moved to useShowData
